--- bsx3/ui/_helpers/auth-header.js
@@ -1,12 +1,0 @@
-
-
-// eslint-disable-next-line import/prefer-default-export
-export function authHeader() {
-  // return authorization header with jwt token
-  const user = JSON.parse(localStorage.getItem('user'));
-
-  if (user && user.token) {
-    return { Authorization: `Bearer ${user.token}` };
-  }
-  return {};
-}
