--- bsx3/ui/actions/alert.actions.js
@@ -1,21 +1,0 @@
-/* eslint-disable no-use-before-define */
-import { alertConstants } from '../constants';
-
-// eslint-disable-next-line import/prefer-default-export
-export const alertActions = {
-  success,
-  error,
-  clear
-};
-
-function success(message) {
-  return { type: alertConstants.SUCCESS, message };
-}
-
-function error(message) {
-  return { type: alertConstants.ERROR, message };
-}
-
-function clear() {
-  return { type: alertConstants.CLEAR };
-}
