+++ bsx3/ui/components/LabeledValue/LabeledValue.jsx
