+++ bsx3/ui/components/LabeledValue/style.css
