+++ bsx3/ui/components/OnOffSwitch/style.css
@@ -1,0 +1,12 @@
+/*  Sytle sheet for OnOffSwitch.jsx */
+
+.inout-switch-msg {
+    text-align: center;
+    float: left
+}
+
+.inout-label {
+    float: left;
+    font-weight: bold;
+    margin-right: 1em
+}
