--- bsx3/ui/components/dataCollection/OnOffSwitch/style.css
@@ -1,3 +1,0 @@
-/*  Sytle sheet for OnOffSwitch.jsx */
-
-.all {}
