+++ bsx3/ui/datacollection/input.css
