+++ bsx3/ui/header/header.css
@@ -4,7 +4,7 @@
         background: -webkit-linear-gradient(to right, #A5CC82, #00467F);
         background: linear-gradient(to right, #A5CC82, #00467F);
         height: 40px;
-        font-weight: bold;
+        /* font-weight: bold; */
     }
     .brand {
         color: rgb(255, 255, 255)!important;
@@ -22,7 +22,7 @@
         margin: 0;
         padding-left: 4px;
         padding-right: 4px;
-        font-weight: bold;
+        /* font-weight: bold; */
     }
     .item1:hover {
         color: rgb(255, 255, 255)!important;
