+++ bsx3/ui/index.jsx
@@ -2,20 +2,13 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'connected-react-router';
-import store, { history } from './store/redux-store';
-import Main from './components/Main';
+import store, { history } from './app/redux-store';
+import Main from './app/Main';
 import * as serviceWorker from './serviceWorker';
-
-// import { initCounterRequest } from './counter/counter-api';
 
 import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-// store.dispatch(initCounterRequest());
-// setup fake backend
-import { configureFakeBackend } from './_helpers/fake-backend';
-
-configureFakeBackend();
 const target = document.querySelector('#root');
 
 render(
